Split webpack configurations into separate files. Add support to extract bundles for vendor files
