+++ webpack/helpers.js
@@ -1,0 +1,11 @@
+const webpack = require('webpack');
+
+const extractBundles = bundles => ({
+  plugins: bundles.map(bundle => (
+    new webpack.optimize.CommonsChunkPlugin(bundle)
+  )),
+});
+
+module.exports = {
+  extractBundles,
+};
